requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asyncer==0.0.8
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
bs4==0.0.2
cachetools==6.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cohere==5.16.1
colorama==0.4.6
coloredlogs==15.0.1
colorgram.py==1.2.0
cryptography==42.0.8
cssselect==1.3.0
cssselect2==0.8.0
cssutils==2.11.1
distro==1.9.0
Django==5.0.6
et-xmlfile==1.1.0
fastapi==0.116.1
fastavro==1.11.1
ffmpy==0.6.0
filelock==3.18.0
filetype==1.2.0
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.4.1
fsspec==2025.5.1
gradio==5.36.2
gradio_client==1.10.4
greenlet==3.2.3
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.7
imageio==2.37.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
llvmlite==0.44.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.5
numba==0.61.2
numpy==2.0.0
onnxruntime==1.22.1
openai==1.95.1
opencv-python-headless==4.12.0.88
openpyxl==3.1.3
orjson==3.10.18
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.3.8
playwright==1.53.0
pooch==1.8.2
premailer==3.10.0
prettytable==3.10.0
protobuf==4.25.4
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pydyf==0.11.0
pyee==13.0.0
Pygments==2.19.2
PyMatting==1.1.14
pyperclip==1.9.0
pyphen==0.17.2
pyreadline3==3.5.4
pyseto==1.7.8
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
referencing==0.36.2
rembg==2.0.67
replit==4.0.0
reportlab==4.4.2
requests==2.32.3
rich==14.0.0
rpds-py==0.26.0
ruff==0.12.3
safehttpx==0.1.6
scikit-image==0.25.2
scipy==1.16.0
selenium==4.23.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
starlette==0.47.1
sympy==1.14.0
tifffile==2025.6.11
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.2
tomlkit==0.13.3
tqdm==4.67.1
trio==0.26.0
trio-websocket==0.11.1
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.3
wsproto==1.2.0
yagmail==0.15.293
yarl==1.9.4
zopfli==0.2.3.post1
